Moved `ItemEchoShard` out of `randomitem` as it's not related to random item selection.
