+++ src/main/java/cn/nukkit/item/ItemEchoShard.java
@@ -1,9 +1,8 @@
-package cn.nukkit.item.randomitem;
+package cn.nukkit.item;
 
 
 import cn.nukkit.api.PowerNukkitXOnly;
 import cn.nukkit.api.Since;
-import cn.nukkit.item.Item;
 
 @PowerNukkitXOnly
 @Since("1.19.50-r1")
